장바구니 API 오류 추가 세분화 하였습니다. 3계층 구조 분리 전 마지막 커밋입니다.

+++ src/service/cart-service.js
@@ -8,7 +8,7 @@ const postCart = async (userIdKey, cartAdd)=> {
         //맨 먼저 유저가 제대로 접속해 있는 상황인지 확인합니다.
         if(!userIdKey){
             throw new Error('먼저 로그인 상태를 확인해 주세요');
-        }
+        };
         //추가하려는 물품의 이름, 가격, 회사, 구매수량을 추가합니다.
         const { name, price, company, quantity } = cartAdd;
 
@@ -24,14 +24,14 @@ const postCart = async (userIdKey, cartAdd)=> {
         const duplicateCart = await Cart.findOne({ name, price, company, poster: userIdKey });
             if(duplicateCart){
                 throw new Error ('이미 장바구니에 추가된 상품입니다.');
-            }
+            };
         //장바구니에 저장합니다.
         await newCart.save();
         //장바구니에 제대로 저장이 되었는지 확인합니다.
         const checkCart = await Cart.find({ name, price, company, poster: userIdKey });
             if(checkCart.length === 0){
                 throw new Error ('현재 장바구니에 상품이 추가되지 않습니다.');
-            }        
+            };    
         //장바구니 document의 ObjectId를 User Schema의 cart 필드에 추가합니다.
             const insertCart = await User.findByIdAndUpdate(
                 userIdKey,
@@ -46,7 +46,7 @@ const postCart = async (userIdKey, cartAdd)=> {
                     지워줘야 합니다.*/
                     await Cart.findOneAndDelete({ _id: newCart._id })
                     throw new Error ('장바구니 추가 중 문제가 발생하였습니다.');
-                }
+                };
             return insertCart.cart;
         } catch (err) {
             throw new Error (`${err.message}`);
@@ -60,13 +60,13 @@ const presentCart = async (userIdKey)=> {
         //맨 먼저 유저가 제대로 접속해 있는 상황인지 확인합니다.
         if(!userIdKey){
             throw new Error('먼저 로그인 상태를 확인해 주세요');
-        }
+        };
         //해당 유저의 모든 장바구니 물품들의 내역을 표시합니다.
         const cartItems = await User.findById(userIdKey).populate('cart');
         //장바구니에 상품이 없다면 오류를 출력합니다.
         if (cartItems.cart.length === 0){
             throw new Error ('현재 장바구니에 상품이 없습니다.');
-        }
+        };
         //장바구니에 담긴 상품들을 나열합니다.
         const arrangeCart = cartItems.cart;
         const cartItemData = arrangeCart.map((item)=> ({
@@ -78,7 +78,7 @@ const presentCart = async (userIdKey)=> {
         }));
         if(cartItemData.length === 0){
             throw new Error ('장바구니에 상품이 출력되지 않았습니다.')
-        }
+        };
         return cartItemData;
     } catch (err) {
         throw new Error (`${err.message}`)
@@ -98,19 +98,25 @@ const removeCart = async (userIdKey, cartId)=> {
         //맨 먼저 유저가 제대로 접속해 있는 상황인지 확인합니다.
         if(!userIdKey){
             throw new Error('먼저 로그인 상태를 확인해 주세요');
-    }
+    };
         //해당 유저의 물품들의 내역을 표시합니다.
         const usersCart = await User.findById(userIdKey).populate('cart');
-        //장바구니에 상품이 없다면 오류를 뱉어냅니다.
+        //장바구니에 상품이 없다면 오류를 출력합니다.
         if (usersCart.cart.length === 0){
             throw new Error ('장바구니에 담긴 물품이 없습니다.');
-        }
+        };
         //유저의 장바구니 품목을 삭제합니다.
         await User.findByIdAndUpdate(
             userIdKey,
             { $pull: { cart: newCartId } },
             { new: true }
-        )
+        );
+        //장바구니의 품목이 삭제되었는지 확인합니다.
+        const isThisCartDeleted = await User.find({ cart: newCartId });
+        console.log(isThisCartDeleted);
+        if(isThisCartDeleted.length !== 0) {
+            throw new Error ('장바구니에 상품이 지워지지 않았습니다.')
+        };
         //참조되고 있던 cart document도 같이 삭제합니다.
         await Cart.deleteOne({ _id: newCartId });
     } catch (err){
@@ -135,10 +141,10 @@ const removeAllCart = async (userIdKey)=> {
         //장바구니에서 모든 상품을 삭제합니다.
         await User.findByIdAndUpdate(userIdKey, { $set: { cart: [] } });
         //만약 제대로 삭제되지 않았다면 오류를 출력합니다.
-        const usersCartUpdate = await User.findById(userIdKey).populate('cart');
-        if(usersCartUpdate.cart.length !== 0){
+        const isUsersCartDeleted = await User.findById(userIdKey).populate('cart');
+        if(isUsersCartDeleted.cart.length !== 0){
             throw new Error ('장바구니 상품이 삭제되지 않았습니다.')
-        }
+        };
         //참조되고 있던 cart document도 전부 삭제합니다.
         const cartIds = usersCart.cart.map((item)=> item._id);
         await Cart.deleteMany({ _id: { $in: cartIds } });
